feat(tier4_autoware_utils): add toHexString
